Auto_Question_AI/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose
passlib[bcrypt]
reportlab
python-multipart
requests
flask
pandas
python-docx
langchain
langchain-groq
python-dotenv
pypdf
faiss-cpu
sentence-transformers
